move splitter uploader to new location
